backend/requirements.txt
Django
djangorestframework
django-cors-headers
mysqlclient
python-dotenv
openai
Pillow
python-jose
django-filter
drf-yasg
gunicorn
whitenoise
django-storages
boto3
celery
redis
django-celery-beat
django-celery-results 